Add config option to keep build env after build.

+++ openshift-environment-driver/src/main/java/org/jboss/pnc/environment/openshift/OpenshiftStartedEnvironment.java
@@ -240,7 +240,9 @@ public class OpenshiftStartedEnvironment implements StartedEnvironment {
     public void destroyEnvironment() {
         client.delete(route);
         client.delete(service);
-        client.delete(pod);
+        if (!environmentConfiguration.getKeepBuildAgentInstance()) {
+            client.delete(pod);
+        }
     }
 
     private String replaceConfigurationVariables(String podConfiguration, Map runtimeProperties) {
